Allows to overwrite current typing.
There is no need to remove and recreate the element in the reset
function. That would break if the span had a class and no id and adds
over complication. This way it is cleaner and bindings are still alive
thus allowing to run the animation again after updating the srtings
values.