Toggle backspace on/off
This change will leave an option to user for toggling backspacing functionality. Not everyone wants their users to wait and watch a bunch of strings rolling back ... when there are a bunch of strings and long sentences to roll back.

I am too much swamped with tasks to checkout a branch in local dev. env. (doing this pull from a web), as I currently don't have node.js on this machine, so I'm sorry for not doing the legit pull request with a gulp build (hence no minified typed.js of this proposed change) ... this is a small change so it can at least serve as a hint to someone else.