Fix HTML element movement caused by empty tag
When typed.js goes from one sentence to the other,
it first does an .empty() on the DOM node,
after which it proceeds to write the sentence
one character at a time.

At .empty(), the corresponding HTML element's
height becomes 0, which in some cases
causes the elements below it to jump up.
When the first character is inserted into
the HTML, the height becomes non-zero again
which causes the elements below to move down.

By replacing .empty() with .html('&nbsp;'),
we ensure that the element's height never
goes down to 0.